refactor/Combine previous, next, detail postAPI
